refactor: apply export and import consistency inside components folder
