+++ src/components/index.ts
@@ -1,0 +1,9 @@
+export { Button } from "./button";
+export { Calendar } from "./calendar";
+export { Contributor } from "./contributor";
+export { CopyrightFooter } from "./copyrightFooter";
+export { Footer } from "./footer";
+export { Header } from "./header";
+export { Logo } from "./logo";
+export { MenuMobile } from "./menuMobile";
+export { SpecialTitle } from "./specialTitle";
