+++ src/components/logo/index.tsx
@@ -4,7 +4,7 @@ import { Wrapper } from "./style";
 import communityLogo from "@/assets/comunityLogoText.png";
 import communityIcon from "@/assets/logo@4x.png";
 
-export default function Logo() {
+export function Logo() {
   const IMAGES = [
     { id: 0, src: communityIcon, alt: "communityIcon", width: 38, height: 32 },
     { id: 1, src: communityLogo, alt: "communityLogo", width: 170, height: 32 },
